restructuring plot WIP + dendrogram WIP + tree property tracking with pyramida-cell
