+++ src/pymodule/dense/_pygrowth.pyx
@@ -1875,10 +1875,10 @@ def _get_tree(neuron, neurite):
     while keep_going:
         keep_going = walk_neurite_tree_(neuron, cneurite, nprop)
 
-        node      = Node(nprop.n_id, tree, parent=nprop.p_id,
-                         diameter=nprop.diameter,
-                         dist_to_parent=nprop.dist_to_parent,
-                         pos=tuple(nprop.position))
+        node = Node(nprop.n_id, tree, parent=nprop.p_id,
+                    diameter=nprop.diameter,
+                    dist_to_parent=nprop.dist_to_parent,
+                    pos=tuple(nprop.position))
 
         tree[int(node)] = node
